requirements.txt
Flask
gunicorn
flask-cors
python-dotenv
openai